refactor native config
