+++ packages/eas-cli/src/update/ios/UpdatesModule.ts
@@ -1,11 +1,11 @@
 import { ExpoConfig } from '@expo/config';
 import { IOSConfig } from '@expo/config-plugins';
 
+import { ensureValidVersions } from '../../build/utils/updates';
 import { getProjectAccountName } from '../../project/projectUtils';
 import { ensureLoggedInAsync } from '../../user/actions';
 import { readPlistAsync, writePlistAsync } from '../../utils/plist';
 import { getVcsClient } from '../../vcs';
-import { ensureValidVersions } from '../utils/updates';
 
 export async function syncUpdatesConfigurationAsync(
   projectDir: string,
